requirements.txt
blinker==1.6.2
click==8.1.6
colorama==0.4.6
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.2
opencv-contrib-python==4.8.0.76
opencv-python-headless==4.8.0.76
Pillow==10.0.0
Werkzeug==2.3.6
